[Major refactoring] Split long script in parts
